git-server: Move function where it's used
