+++ git-server/src/hooks/pre_receive.rs
@@ -16,12 +16,14 @@
 //!
 //! The `pre-receive` git hook provides access to GPG certificates for a signed push, useful for authorizing an
 //! update the repository.
+use std::io;
 use std::io::prelude::*;
 use std::io::stdin;
 use std::str::FromStr;
 
 use envconfig::Envconfig;
 use git2::{Oid, Repository};
+use librad::PeerId;
 
 use super::{
     types::{CertNonceStatus, CertStatus, ReceivePackEnv},
@@ -122,7 +124,7 @@ impl PreReceive {
             // key fingerpint.
             let (peer_id, _) = crate::parse_ref(refname)
                 .map_err(|_| Error::InvalidRefPushed(refname.to_owned()))?;
-            let peer_fingerprint = crate::to_ssh_fingerprint(&peer_id)?;
+            let peer_fingerprint = to_ssh_fingerprint(&peer_id)?;
 
             if key_fingerprint[..] != peer_fingerprint[..] {
                 return Err(Error::Unauthorized("signer does not match remote ref"));
@@ -183,4 +185,21 @@ impl PreReceive {
 
         Err(Error::Unauthorized("key is not authorized to push"))
     }
+}
+
+/// Get the SSH key fingerprint from a peer id.
+fn to_ssh_fingerprint(peer_id: &PeerId) -> Result<Vec<u8>, io::Error> {
+    use byteorder::{BigEndian, WriteBytesExt};
+    use sha2::Digest;
+
+    let mut buf = Vec::new();
+    let name = b"ssh-ed25519";
+    let key = peer_id.as_public_key().as_ref();
+
+    buf.write_u32::<BigEndian>(name.len() as u32)?;
+    buf.extend_from_slice(name);
+    buf.write_u32::<BigEndian>(key.len() as u32)?;
+    buf.extend_from_slice(key);
+
+    Ok(sha2::Sha256::digest(&buf).to_vec())
 }
